try to subclass rctvideo
